jexport: implement -d:jnimGenDex
This change makes it possible to use `jexport` together with `jnimDexWrite` to
emit .dex files. The `jnimDexWrite` proc builds a .nim file, which can then be
compiled using package dali to generate a .dex file.

Currently, the intermediate step of generating a .nim file is required, as
Nim's compile-time has a lot of limitations. In this particular case, see e.g.:
- https://github.com/nim-lang/Nim/issues/14339
- https://github.com/nim-lang/Nim/issues/14340